refactor: model 폴더에 타입 관리 파일 생성
