+++ src/entities/model/informationModel.ts
@@ -1,0 +1,28 @@
+export type GatheringsInfoTypes = {
+  teamId: number;
+  id: number;
+  type: string;
+  name: string;
+  dateTime: string;
+  registrationEnd: string;
+  location: string;
+  participantCount: number;
+  capacity: number;
+  image: string;
+  createdBy: number;
+  canceledAt: string;
+};
+
+export type ParticipantListTypes = {
+  teamId: number;
+  userId: number;
+  gatheringId: number;
+  joinedAt: string;
+  User: {
+    id: number;
+    email: string;
+    name: string;
+    companyName: string;
+    image: string;
+  };
+};
